+++ docsrc/conf.py
@@ -11,5 +11,5 @@ repo_name = "publish"
 
 setup_globals(org, org_name, repo, repo_name)
 
-html_theme = "foundation"
+